src/AWS-IoT-Grenngrass-Components/artifacts/com.example.image_items/1.0.0/requirements.txt
opencv-python==4.5.3.56
roboflow
awsiotsdk